+++ Tools/Colorizer/Template/Defaults.js
@@ -1,0 +1,33 @@
+
+export default {
+
+    Foreground : {
+        Color : '090d11' ,
+        Alpha : .75
+    },
+
+    Background : {
+        Color : 'ffffff' ,
+        Alpha : .5
+    },
+
+    Accent : {
+        Color : '2598e4' ,
+        Alpha : 1
+    },
+    
+    Success : {
+        Color : '42a53b' ,
+        Alpha : 1
+    },
+
+    Warning : {
+        Color : 'f2712c' ,
+        Alpha : 1
+    },
+        
+    Error : {
+        Color : 'f03489' ,
+        Alpha : 1
+    }
+}
